+++ src/main/java/com/example/casestudymd3_shopxanh/model/Category.java
@@ -1,7 +1,7 @@
 package com.example.casestudymd3_shopxanh.model;
 
 public class Category {
-    private Long id;
+    private int id;
     private String name;
 
     public Category(String name) {
@@ -11,16 +11,16 @@ public class Category {
     public Category() {
     }
 
-    public Category(Long id, String name) {
+    public Category(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
