+++ src/main/java/com/example/casestudymd3_shopxanh/model/Role.java
@@ -1,22 +1,22 @@
 package com.example.casestudymd3_shopxanh.model;
 
 public class Role {
-    private Long id;
+    private int id;
     private String name;
 
     public Role() {
     }
 
-    public Role(Long id, String name) {
+    public Role(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
