+++ src/main/java/com/example/casestudymd3_shopxanh/service/IGeneralService.java
@@ -5,9 +5,9 @@ import java.util.*;
 
 public interface IGeneralService<T> {
     void add(T t) throws SQLException;
-    T findById(Long id) throws SQLException;
+    T findById(int id) throws SQLException;
     List<T> findAll();
     List<T> findByName(String name);
-    boolean delete(Long id) throws SQLException;
+    boolean delete(int id) throws SQLException;
     boolean update(T t) throws SQLException;
 }
